Updated path to renamed directory
